Move ILayerOption to map types
