+++ src/components/Map/Layers/Buildings/HelsinkiBuildings/types.ts
@@ -1,18 +1,11 @@
 import { StringValueOf } from './general'
+import { ILayerOption } from '#/common/types/map'
 
 export enum LayerLevel {
-    Helsinki_buildings = 'fi_miscs_helsinki_buildings',
+  Helsinki_buildings = 'fi_misc_helsinki_buildings',
 }
 export type LayerLevelStrings = StringValueOf<LayerLevel>
 
-export interface ILayerOption {    
-    serverId: string    
-    minzoom: number
-    maxzoom?: number
-    layerMinzoom?: number | null
-    layerMaxzoom?: number | null
-}
-  
 export type ILayerOptions = {
-    [key in LayerLevelStrings]: ILayerOption
+  [key in LayerLevelStrings]: ILayerOption
 }
